+++ components/SEO.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import LocalBusiness from './SEO/LocalBusiness'
+import WebSite from './SEO/WebSite'
+import Service from './SEO/Service'
+
+const SEO = () => (
+	<React.Fragment>
+		<WebSite />
+		<LocalBusiness />
+		<Service />
+	</React.Fragment>
+)
+
+export default SEO