+++ pages/_document.js
@@ -1,6 +1,8 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document'
 import { ServerStyleSheet } from 'styled-components'
 import { registerServiceWorker } from '../helpers/utils'
+import SEO from '../components/SEO'
+import GTM from '../components/GTM'
 
 const env = {
 	TIER: process.env.TIER,
@@ -54,6 +56,8 @@ class MyDocument extends Document {
 						href="/images/icons/crafted-jewellers-192.png"
 					/>
 					{registerServiceWorker()}
+					<GTM />
+					<SEO />
 				</Head>
 				<body>
 					<noscript>